backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oracledb==3.0.0
packaging==24.2
pycparser==2.22
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
python-dotenv==1.0.0